addressed review comments.
