+++ cli/src/core/ast/infrastructure.rs
@@ -1,4 +1,5 @@
 pub(crate) mod database;
+pub(crate) mod env;
 pub(crate) mod kafka;
 pub(crate) mod redis;
 pub(crate) mod s3;
